+++ tests/testthat/libtest.R
@@ -30,6 +30,9 @@ test_raw_locality <- function(locality) {
     expect_equal(names(locality), c("metadata", "loggers"))
     test_locality_metadata(locality$metadata)
     expect_equal(class(locality$loggers), "list")
+    expect_false(is.null(names(locality$loggers)))
+    expect_false(any(duplicated(names(locality$loggers))))
+    expect_equal(names(locality$loggers), unname(purrr::map_chr(locality$loggers, ~ .x$metadata@name)))
     purrr::walk(locality$loggers, test_logger)
 }